+++ src/Foundation/Providers/ModuleServiceProvider.php
@@ -29,11 +29,13 @@ use Vanilo\Foundation\Factories\OrderFactory;
 use Vanilo\Foundation\Models\Address;
 use Vanilo\Foundation\Models\Cart;
 use Vanilo\Foundation\Models\Customer;
+use Vanilo\Foundation\Models\MasterProduct;
 use Vanilo\Foundation\Models\MasterProductVariant;
 use Vanilo\Foundation\Models\Order;
 use Vanilo\Foundation\Models\Product;
 use Vanilo\Foundation\Models\Taxon;
 use Vanilo\Foundation\Models\Taxonomy;
+use Vanilo\MasterProduct\Contracts\MasterProduct as MasterProductContract;
 use Vanilo\MasterProduct\Contracts\MasterProductVariant as MasterProductVariantContract;
 use Vanilo\Order\Contracts\Order as OrderContract;
 use Vanilo\Order\Contracts\OrderFactory as OrderFactoryContract;
@@ -63,6 +65,7 @@ class ModuleServiceProvider extends BaseBoxServiceProvider
         $this->concord->registerModel(TaxonomyContract::class, Taxonomy::class, $registerRouteModels);
         $this->concord->registerModel(CartContract::class, Cart::class, $registerRouteModels);
         $this->concord->registerModel(OrderContract::class, Order::class, $registerRouteModels);
+        $this->concord->registerModel(MasterProductContract::class, MasterProduct::class, $registerRouteModels);
         $this->concord->registerModel(MasterProductVariantContract::class, MasterProductVariant::class, $registerRouteModels);
 
         Relation::morphMap([